Changed hash function vectorization to accomodate hash offsets.
This requires repeating the table_size, and then passing the table_size,
hash_offset, and vertex into the hash function all with shape[0] ==
num_levels. I think they're might be a more efficient way to do this
that doesn't require repeating the table size into a vector of the
correct shape. Seems kind of wasteful since it will be constant
throughout training. Maybe repeat it once when the table is initialized?
That still requires passing the duplicated data around though. I could
also wrap the hash function in a class with table_size as a member, but
I would prefer to stick with a functional approach. Maybe if I grok
jax.vmap I can figure out how to vectorize the vertex and hash_offset
without vectorizing the table_size.