Now calculating ray deltas with origins.
Adding ray origins to the ray samples before deltas are calculated means we get n deltas instead of n-1, where n is the number of ray samples.

Without the origins, deltas would be calculated starting at the first
and second ray sample, this means there is no delta describing how far
along the first ray sample is. Adding ray origins to the beginning of
the ray samples means that the deltas are instead calculated starting at
the origin and first sample. This leads to a total of n samples.