Turned ray near into function argument.
Ray near is now passed as an argument to train_loop and render_scene.
Because of this, these functions no longer add ray near to ray far when
calculating ray scales. Now both of these parameters are set manually in
the __main__ section of the script. There is also now an assertion to
make sure ray near is less than ray far. This should avoid the ray scale
and division by zero error that I fixed in the last couple of commits,
while also providing more control of the script's behaviour.