Added original canvas rays to scaled rays.
When rays are normalized, they recede from the canvas plane, this
effectively turns the near plane into a partial unit sphere. By adding
the original canvas rays (which all point to the canvas plane) to the
normalized rays, after the normalized rays have been scaled, we get a
near plane which is once again equal to the canvas plane. This is
because the scaled rays will now all begin at a point on the canvas and
move outwards away from the center of projection, therefore they won't
detect any density which lies before the canvas plane.