Reimplemented random background colors for training.
This should help ensure that the network is learning the correct
densities, because densities determine the alpha composition of the
rendered pixel, so they must be correct in order for the rendered pixel
to match the source pixel when the random backgrounds are alpha
composited onto them. Because the background is random, the network
cannot simply memorize its color in order to avoid proper alpha
composition.