Seperately normalized rotation and translation transformations.
As part of the dataset preprocessing, I extracted the rotation and
translation components of the transformation matrices and normalized
them seperately. Then I recombined them to get the complete transform
matrices. This ensures that all camera positions are on the unit sphere
and makes it easy to keep rays mostly within the unit cube.