Sequential batch ray marching that returns compacted buffer.
The idea here is to save valid samples to a tightly packed buffer while
discarding the others, then store where each ray's samples begin and end
in a second buffer. When rendering, all of a ray's valid samples can be
retrieved by taking a slice of the tightly packed buffer from the desired ray's start position to the next ray's start position.