Added random background colors for training.
This makes the model start to diverge in fewer steps, so I think it is a
step in the right direction. The reason being that the model gets to the
point where the loss landscape is narrow enough launch the loss back up
more quickly, thus it learns generalities more quickly.

One thing I'm not sure about here is if the gradient calculation is
still correct now that the rendered_alpha is directly affecting the
loss. I'll need to look further into this.