Switched to constant z_step with box bounded ray samples when tracing.
Previously the number of ray samples was bounded by a hyperparameter
which determined the number of z_vals in the linear space between ray
near and ray far. Only the z_vals from that linear space which produced
ray samples in the [0, 1] unit box were considered valid samples, so in
practice the number of ray samples was usually much lower than the upper
bound. Now the z_vals for any given ray are computed by taking a
series of constant z_steps along the ray, checking if the ray
at that step is in the [0, 1] unit box, and if it is then that z_val/ray
position is a valid sample. Each ray is given a maximum number of
samples that it is allowed to generate. In practice this is the
difference of the batch size and the number of samples that have already
been accumulated. Therefore, with a small enough step size, a single ray
could theoretically generate enough samples for the entire batch if the
number of accumulated samples is 0 when that ray begins tracing. This is
a prerequisite to the implementation of an occupancy grid since some
rays will go through more densely populated spaces than others, and thus
produce more samples. This change also drastically improved the quality
of renders from novel views. I believe this is because the constant
z_step prevents the distance between samples from stretching out as the
horizontal and vertical angle of the ray increases. This stretching
occurs because the distance along a ray from the near plane to the far
plane at larger angles is greater than at smaller angles, thus each step
in the linear space must be larger at larger angles. There might be
other reasons for this drastic improvement, but that's my best guess
right now.