Added argument piping from synthesize_data to blender proc gen script.
This allows for control of the number of renders and the output
directory of those renders when calling the top level data synthesis
script. It should also enable us to run multiple instances of the data
synthesis script in parallel and have them output to different
directories so they don't overwrite each other.