Fixed blender download with wget.
I don't think it's possible to cache the blender download to a folder on
the host machine, so instead there should be a script that runs when the
container is started with "docker-compose up" that checks if blender has
been cached, and if not caches the installed version. This will enable
writing to folders on the host since the docker-compose file mounts
them.