Added scene rendering to vrj test.
It doesn't seem to be working very well right now. I'm not sure if the
cause of this is the scene renderer itself or the model. Jaxngp uses
different ray marching and integration functions for inference, so that
could be the problem. On the other hand, the model is learning a strangely
shaped density function which can be seen when visualizing the occupancy
grid density as well as the regular density grid, so this could also be
the problem.