Debbuging strange density function shape.
The density function, which I'm inspecting via the occupancy grid mask,
has a strange shape. With some training configurations it vaguely
resembles the shape that it's supposed to (i.e. the 3D reconstruction of
the dataset), but it's very rounded/bulbous. Like it was filled with air
and lost its shape. I think this means that the model is
underconstrained, but I don't understand how it could be.