Set user and switched entrypoint to infinite loop.
The user is set using the environment variables $UID and $GID. Each user will
have to set these values themselves. On linux this can be accomplished with:

export UID=$(id -u)
export GID=$(id -g)

I also switched the entrpoint to an infinite loop because jupyterlab was
having trouble starting when the container was run as a user instead of
root. This is fine since I wasn't really using jupyterlab anymore.

The reason for this change was that files created inside the container
were owned by root so I want able to move them around on my host
without sudo.