Added untested generated weight rendering for sliding ltd.
Untested because I'm currently waiting for an unpacked nerf dataset to
finish transferring from an external drive so I can pack it and get the
weight map file. This file is required to unpack the generated weights
for rendering, so I won't know how well rendering works until I get it.