Added script to convert nerf dicts to new module structure.
The main difference is that the dense layers are now grouped into feed
forward modules, so Dense_N becomes Dense_P when grouped under
FeedForward_M, where N is the index of the dense layer in the old
structure, M is the index of the feed forward layer in the new structure,
and P is the index of the dense layer in the new structure relative to
its parent feed forward layer. The script isn't general and
there is a very specific hardcoded mapping. Dense_0 and Dense_1 now fall
under FeedForward_0, while Dense_2, Dense_3, and Dense_4 now fall under
FeedForward_1. Their indices are changed accordingly.