Prototyped context squashing calculations.
The idea is to squash all the contexts required to cover a sequence so
that they overlap and have minimal padding. So if 1000 padding
tokens are required for naive context splitting, then these tokens will
be evenly distributed into overlapping sections of adjacent contexts.
This has the effect of "squashing" the contexts down into the sequence.
After this the number of required padding tokens will be the remainder
that could not be evenly distributed. Note that the First context is
anchored to the start of the sequence, so it cannot be squashed.
Consequently the context overlap is:

naive_padding_size // (num_contexts - 1)

where the -1 accounts for the unmoveable first context.