Moved top level data synthesis script into synthdata as __main__.
This script serves as a general entrypoint for data synthesis and is
responsible for calling the blender scripts which actually generate the
data. Now instead of calling scripts/synthesize_data.py (which has been
deleted), use the following command:

python3 -m synthdata <arguments here>

This change is part of the reorganization for issue #6.