Generated weight clipping.
Generated weights have a very large maximums (30-200) and very small minimums (negative 30-200).
In contrast, the source weights are usually in the range of [-1.5, 1.5].
This disparity seems to be caused by the signal rate division in reverse
diffusion. When the denoised batch is divided by the signal rate (which
is very small at the start of reverse diffusion), it becomes very large.
For example, even if the denoised batch has a range of [-5, 5], when it
is divided by a signal rate of 0.02, it becomes [-250, 250]. I think
this is supposed to be balanced over time by the falling noise rate, but
I'm not sure. This didn't seem to be a problem with the NTG
implementation so I'll run some tests in that repo to see if it
demonstrates the same behaviour. Anyways, for now I'm clipping the
generated weights to [-2, 2] so that the renders will only show detail
from weights that are mostly in the correct range.