Added redundancy check and PIL image close for generate_ngp_images.
For some reason the script gets killed after about a thousand
iterations, in order to pickup where it left off without redoing work, I
added a redundancy check. This checks whether a file corresponding to
the input file already exists in the output directory, and skips that
input file if there is. I also added a line to close the PIL image at
the end of each iteration because I suspect python might not be garbage
collecting it, thus causing the script to be killed when it runs out of
memory.