Fixed checkpoint epoch index.
Previously checkpoints were loaded if the checkpoint_epoch arg was
greater than 0. This prevented the checkpoint for epoch 0 from ever being
loaded. Now the default checkpoint_epoch arg is -1, and checkpoints are
loaded if it is greater than -1. The train loop now also starts at
epoch = checkpoint_epoch + 1, so loading the checkpoint for epoch 0 will
cause training to start at epoch 1, and so on. Note that the default
value of -1 for checkpoint_epoch will cause training to start at epoch
0.