LADiT training with smaller token sizes.
Previously the token size was fixed to a field's MLP width since this is
the way the weights are packed. The rationale for using smaller token
sizes is to lean more on the attention function and less on the dense
layers while also allowing the attention function to be more precise
(since each token now represents fewer weights).