Added transformer VAE to hypernets.
I think the LADiT has failed to generate valid neural fields because of
their sparsity in weight space. To account for this, I'm going to try to
encode the weights into a smoother and denser latent space using a VAE,
before sending them to the LADiT. This is basically latent diffusion but
with weights instead of images.