Switched reconstruction loss from MSE to binary cross entropy.
I read that the reconstruction loss is supposed to be summed instead of
averaged, but I'm not sure if this applies to the batch dimension as
well. When I sum the sample loss and then average across the batch, the
test generations are bad. When I sum the entire sample/batch loss, the
loss is very high, but the test generations are good. Maybe sum is the
correct way but it's necessary to scale the total loss?