Switched vanilla diffusion norms/softmax layers to use the custom dtype.
Previously these were hardcoded to fp32 due to fears of numerical
instability, but I did some tests with bf16 and they seem to be stable.
Some additional configuration logic may be required in order to support
fp16 as I'm not sure whether or not these layers will still be stable
with that dtype.