Patched IF DiT scripts to support token_dim > 1 by reshaping samples.
Now all samples are reflattend before being unpacked and rendered as
fields. In practice this just means taking them from (n, c, t) to
(n, c * t) where n is the number of samples, c is the context length,
and t is the token dim. This transformation is necessary for the
unflattening code to work.