Fixed index parsing bug.
The middle index group was being skipped over because when
current_char == ' ' and index_group_start != -1 at the end of the first
group it reset index_group_start to -1 and moved onto the next
character, but the next character was an index instead of a space, so
the condition to start the next group was not triggered until reaching
the next space where index_group_start == -1 which was the third group.
Automatically starting a new group at the end of the last group (by
setting the group start position and resetting the index positions)
fixes this. Now the first if block only starts the first group and
all others are started at the end of the previous in the final else-if
block.