Minor refactor of obj parser.
Mostly just moved state resets into the end of line blocks for branches
that use that state. For example, the is_face branch will reset is_face
and the parser indices when it gets to the end of the line. This way
states that have not been used to parse the current line, and therefore
have not been changed from default will not be reset unnecessarily.

Also turned buffer range checks into early escapes via return NULL.