Fixed obj normal loading.
Normals are associated with vertices according to the face index groups.
By default the order of normals is not the same as the order of
vertices. There may also be more or less normals than vertices. In order
to have normals and vertices share the same relative position within
their respective buffers, the normals must be reordered according the
the normal/vertex indices.