Deleted jax-tcnn and tiny-cuda-nn.
jax-tcnn relies on tiny-cuda-nn which is a pretty complex dependency
(and very large due to the json lib). The multiresolution hash encoding
(which is the only thing we needed tiny-cuda-nn for) can be
reimplemented as a much simpler cuda dependency if it is needed again in
the future. For now we'll just use the jax implementation.