Removed MTL parsing in favor of directly specifying texture path.
We will likely want to use many different shading styles for each mesh,
so it will be easier to hardcode their properties into the program, or
store them in a json file that is read by python and then sent to the c
program at runtime. Since MTL files are no longer parsed, we specify
directly specify a texture path for each mesh and load it with the
load_png() function. The OBJ struct and the texture struct
are then combined into a single gl mesh by obj_to_gl_mesh().