Viewport dimensions are now sent to render scene function.
Now the gl viewport doesn't resize to a hardcoded 1280x720 after
rendering the depth map, instead it resizes to whatever the provided
viewport dimensions are. In this case those are the values of the global
variables window_width and window_height. This allows window resizing
via the gl callback which isn't very important since the callback will
not be present in the final version, but it's convenient for
development. This also ensures that there is a single source of truth
for the viewport dimensions which is much more important.