Simplified tokenized field vae and removed compression component.
Compressing neural fields into latent vectors with dimensions less than
the original parameter counts has proven challenging with a VAE.
Intuitively this makes sense to me because neural fields are already
compressed representations of the sensor data that they were trained on.
Because of this it might be easier to use the VAE solely for mapping
the parameter tokens into a smooth and continous latent space while
leaving further compression to the more powerful diffusion model.