Switched to saving camera model matrix as transform matrix.
The transform matrix for nerf camera poses needs to transform rays from
world space to world space instead of world space to camera space. This
is accomplished by saving the camera's model matrix instead of its
lookat matrix. Note that the ray direction has to be inverted when
transforming rays with this method, otherwise it will point away from
the origin relative to the camera's position. Poses from the orbit camera
(which is what we're currently using) should always point towards the
origin relative to the camera's position.