Tokens loaded by HF datasets don't seem to match the generated tokens.
Printing out the tokens as they are generated in the
tokenize_field_params script shows that they are within the correct
token 11 bit token range. This means there is either a bug in the
parquet saving or HF is not loading the dataset correctly. Both are
weird because the tokenized mnist dataset was loaded correctly before.