Modified ar_hypernet sampling to remove start token and add last token.
The start token is just a place holder, it needs to be removed so that
the last token can be added whilst maintaing the fixed context length.
If this is not done then field parameters will not be valid after
detokenization because there is no "start parameter". The first field
paramater corresponds to the second token of the sampled sequence before
the start token is removed.