Modified pre/post latent dense block and removed padding before loss.
Removing padding before the loss is meant to make reconstruction easier
by training only on the parts of the sequence necessary for downstream
application, however it seems to produce this xla warning:

2024-05-28 20:59:30.121217:
W external/xla/xla/service/gpu/gpu_conv_rewriter.cc:372] The low padding
of the backward convolution would be negative (-132), which isn't supported
by GpuConvPaddingLegalization for now (b/32744257).

Despite this the current training run hasn't crashed, so I'm not sure
how serious it is.