Converted even_better_conv_ae back to pure convolutions.
The dense bottleneck layers were causing overfitting even with a
dropout rate (0.7) so I'm switching back to pure convolutions. I also
modified the last block of the encoder and the first block of the
decoder so that they don't downsample/upsample. My is that this
will let the autoencoder setup features in a way that makes them
cleanly "collapsible" by the final convolutional bottleneck without
the burden of a destructive maxpooling layer.