Renamed dependencies to external and created submodules dir.
With this new setup, ./external will contain all code that is developed
outside of this repo and should therefore not be changed except for
copy/pasted whole version updates. Conversely, ./submodules will contain
code that is developed within this repo but that requires a
build/installation step before it can be used. The submodules directory
serves to separate such code from other internally developed code which
doesn't require build/installation steps (i.e. regular scripts).

The direction of dependencies should be:
regular scripts -> submodules -> external

For the nix devshell, external packages are manually imported and then
added to nixpkgs via an overlay so that submodules can access them
during their build. The submodules are automaticall collected into an
overlay by ./submodules/default.nix which is then added to nixpkgs so
that everything is accessible via `with pkgs;` when building the
devshell.