Removed bitfield16 incorrect write to input and fixed output size.
There was some code left over from the fp32_to_bitfield16 kernel for
writing bits with the mask result. The destination for these writes was
mistakenly changed from input to output. The code also shouldn't have
been in the bitfield16_to_fp32 kernel to begin with, so it has been
removed. The output size was also incorrectly set during mlir lowering.
It was being set to batch.shape[0] which is the dimension of the input,
but for the bitfield16_to_fp32 kernel the input dimension must be
divided by 16 to get the output dimension. This has been fixed by
changing the input to the descriptor creation function.

With these changes the test_bitfield() test in
tests/test_fp_tokenization.py now passes.