Casting single and byte pair tokens to uint8 fixes grey renders.
I thought the grey renders from single byte tokens were a precision
issue, but it was actually a data loading/detokenization issue. For some
reason HuggingFace Datasets wasn't storing the tokens as uint8
(instead it was storing them as int32), so the detokenization kernel
which expects uint8 tokens was not working correctly. Converting
tokens to uint8 after they're loaded fixes this and allows for proper
detokenization. When the tokens are detokenized properly the renders
are no longer grey.